added feture to strore the trip details
