+++ theroute/backend/db.js
@@ -1,0 +1,19 @@
+// backend/db.js
+const mysql = require('mysql2');
+
+const db = mysql.createConnection({
+  host: 'localhost',
+  user: 'root',
+  password: 'Nir01620622513', // Your MySQL password
+  database: 'tripdb', // The database created earlier
+});
+
+db.connect((err) => {
+  if (err) {
+    console.error('Error connecting to MySQL:', err);
+  } else {
+    console.log('Connected to MySQL');
+  }
+});
+
+module.exports = db;
